Prevent snapshot/exp utils from being bundled into the js client
